Add files via upload
A common mantra in the NHL is to ‘get pucks on net.’ This leads to the natural question of: ‘ does taking more shots on goal correlate with goals scored by that player or assists by that player?’ Figure 1 shows the distribution of goals (blue circles) and assists (red circles) versus player shots on goal across the 2005-2006 to 20011-2012 NHL seasons. There is an overall positive trend in goals and assists with shots on goal, though it is also clear that shots on goal does not guarantee the most goals or assists. For example, the player with the most shots on goal on the far right has neither the most single season assists or the most single season goals.