Add files via upload
One approach to optimizing a hockey team is to optimize the defensive pairings for that team. Specifically, each hockey team typically has 6 defensemen that play in 3 pairs. In figure 2, I used linear programming to find the best pair combinations that maximize each defensive pairs overall ‘plus/minus’ statistic. In hockey, the ‘plus/minus’ statistic is a cumulative statistic that reflects a players overall defensive and offensive contributions; a positive plus/minus means the player is generally on the ice for more goals for their team than against, while a negative plus/minus means the player is on the ice for more goals against their tam than for their team. Therefore, a defensemen, whose job it is to prevent goals, would ideally have a positive plus/minus statistic. By finding the optimal defense pairings based on plus/minus, this suggests that each pair will combine to be on the ice for the least amount of goals against and most goals for, thereby maximizing their teams chance of winning.

Some interesting points in figure 2 can be observed in the regions delineated by the black and red boxes. First most people would agree that a defensemen that scores 5 or fewer goals in a season would be categorized as a ‘defensive’ defensemen, as opposed to a defensemen who scores 10 or more goals who would be categorized as an ‘offensive’ defensemen. Thus, looking in the black square, it can be observed that defensemen scoring 10 or more goals are often paired with a defensemen scoring 5 fewer goals. This suggests that a team looking to maximize a defensive pair’s plus/minus should pair an ‘offensive’ defensemen with a ‘defensive’ defensemen. In contrast, combining two ‘offensive’ defensemen rarely results in an optimal combination based on defensive pair plus/minus.